# docs/swagger.yaml
definitions:
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
    type: object
...
  dto.Review:
    properties:
      buyerID:
        type: string
      buyerName:
        type: string
      date:
        type: string
      image:
        type: string
      message:
        type: string
      reviewID:
        type: string
      score:
        type: integer
      sellerID:
        type: string
      sellerName:
        type: string
    type: object
  dto.ReviewCreateRequest:
    properties:
      buyerID:
        type: string
      buyerName:
        type: string
      image:
        type: string
      message:
        type: string
      score:
        type: integer
      sellerID:
        type: string
      sellerName:
        type: string
    type: object
  dto.ReviewUpdateRequest:
    properties:
      image:
        type: string
      message:
        type: string
      score:
        type: integer
    type: object
...
  dto.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
    type: object
...
info:
  contact: {}
paths:
...
  /review/:
    get:
      consumes:
      - application/json
      description: Retrieves all reviews
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.Review'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get all reviews
      tags:
      - review
    post:
      consumes:
      - application/json
      description: Creates a new review in the database
      parameters:
      - description: Review to create
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/dto.ReviewCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Review'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new review
      tags:
      - review
  /review/{review_id}:
    delete:
      consumes:
      - application/json
      description: Delete a review's data by its ID
      parameters:
      - description: Review ID
        in: path
        name: review_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete a review by ID
      tags:
      - review
    get:
      consumes:
      - application/json
      description: Retrieves a review's data by its ID
      parameters:
      - description: Review ID
        in: path
        name: review_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Review'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get a review by ID
      tags:
      - review
    put:
      consumes:
      - application/json
      description: Updates an existing review's data by its ID
      parameters:
      - description: Review ID
        in: path
        name: review_id
        required: true
        type: string
      - description: Review data to update
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/dto.ReviewUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Review'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update a review by ID
      tags:
      - review
  /review/buyer/{buyer_id}:
    get:
      consumes:
      - application/json
      description: Retrieves each buyer's reviews by buyer ID
      parameters:
      - description: Buyer ID
        in: path
        name: buyer_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.Review'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get reviews by buyerID
      tags:
      - review
  /review/seller/{seller_id}:
    get:
      consumes:
      - application/json
      description: Retrieves each seller's reviews by seller ID
      parameters:
      - description: Seller ID
        in: path
        name: seller_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.Review'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get reviews by sellerID
      tags:
      - review
...
swagger: "2.0"